1. convertToKrw 내에서 exchangeRateService.fetchAndSaveExchangeRate()를 호출하는 로직을 완전히 제거 2. getDailySettlement와 getTotalSettlement 간에 중복되던 멤버별 정산 계산 로직을 calculateMemberSettlementData 프라이빗 메서드로 추출하여 통합 3. 두 서비스 파일에서 중복되던 환율 변환 로직(processExchange/processExchangeDto)을 제거하고, ExchangeRateProcessor.process()를 호출하도록 변경 4. API DTO를 Currency 엔티티로 변환하는 로직이 정의되었으며, 특히 JPY(100) 통화 코드를 1단위 환율로 정확하게 계산하고 DB 스케일(4자리)에 맞추는 로직이 포함
