+++ src/main/kotlin/com/tribe/tribe_api/exchange/service/ExchangeRateScheduler.kt
@@ -2,16 +2,15 @@ package com.tribe.tribe_api.exchange.service
 
 import com.tribe.tribe_api.common.util.DateUtils
 import com.tribe.tribe_api.exchange.client.ExchangeRateClient
-import com.tribe.tribe_api.exchange.dto.ExchangeRateDto
-import com.tribe.tribe_api.exchange.entity.Currency
+// import com.tribe.tribe_api.exchange.dto.ExchangeRateDto // 더 이상 필요 없음
+// import com.tribe.tribe_api.exchange.entity.Currency // 더 이상 필요 없음
 import com.tribe.tribe_api.exchange.repository.CurrencyRepository
+import com.tribe.tribe_api.exchange.util.ExchangeRateProcessor // ✅ 추가
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.math.BigDecimal
-import java.time.LocalDate
 
 @Service
 class ExchangeRateScheduler(
@@ -40,7 +39,7 @@ class ExchangeRateScheduler(
 
             // 2. 필터링 로직 없이 모든 통화를 처리하도록 수정
             val currenciesToSave = exchanges.mapNotNull { dto ->
-                processExchange(dto, apiDate)
+                ExchangeRateProcessor.process(dto, apiDate) // ✅ 변경: 유틸리티 호출
             }
 
             // 3. DB에 저장/업데이트 (PK가 같으면 UPDATE)
@@ -50,48 +49,5 @@ class ExchangeRateScheduler(
         } catch (e: Exception) {
             log.error("Failed to update exchange rate: {}", e.message, e)
         }
-    }
-
-    /**
-     * 통화 필터링 로직을 제거하고, 100단위 통화만 1단위로 변환하여 Currency 엔티티를 생성합니다.
-     */
-    private fun processExchange(dto: ExchangeRateDto, date: LocalDate): Currency? {
-        val rawCurrencyUnit = dto.curUnit
-
-        // 1. 통화 코드 결정 및 100단위 처리 여부 확인
-        val is100Unit = rawCurrencyUnit.endsWith("(100)")
-        val targetCurrency: String = if (is100Unit) {
-            rawCurrencyUnit.substringBefore('(') // JPY(100) -> JPY
-        } else if (rawCurrencyUnit.contains('(') || rawCurrencyUnit.contains(')')) {
-            return null // 알 수 없는 괄호 형식은 무시
-        } else {
-            rawCurrencyUnit // USD, EUR 등 일반 통화
-        }
-
-        // 한국 원화(KRW)는 KRW/KRW이므로 저장할 필요 없습니다.
-        if (targetCurrency == "KRW") {
-            return null
-        }
-
-        // 2. 쉼표(,) 제거 후 BigDecimal로 파싱
-        var exchangeRate = try {
-            // 매매 기준율(deal_bas_r) 사용
-            BigDecimal(dto.dealBasR.replace(",", ""))
-        } catch (e: NumberFormatException) {
-            log.warn("Failed to parse exchange rate for {}: {}", dto.curUnit, dto.dealBasR)
-            return null
-        }
-
-        // 3. 100단위로 받은 경우 1단위로 변환 (JPY(100) 등)
-        if (is100Unit) {
-            exchangeRate = exchangeRate.divide(BigDecimal("100"))
-        }
-
-        return Currency(
-            curUnit = targetCurrency,
-            curName = dto.curName,
-            exchangeRate = exchangeRate,
-            date = date
-        )
     }
 }
